refactor: extract shared test helpers, fix API key fixture
Address code review feedback:
- Extract captureLogger and mockSpawnProc into src/test-helpers.ts
  to eliminate duplication between connect.test.ts and logs.test.ts
- Replace realistic API key fixture ('nvapi-...') in validate.test.ts
  with 'test-api-key-placeholder' to avoid secret scanner triggers
