+++ nemoclaw/src/commands/connect.test.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { describe, it, expect, vi, beforeEach } from "vitest";
-import type { PluginLogger } from "../index.js";
+import { captureLogger, mockSpawnProc } from "../test-helpers.js";
 
 // ---------------------------------------------------------------------------
 // Mocks
@@ -17,51 +17,6 @@ const { spawn } = await import("node:child_process");
 const { cliConnect } = await import("./connect.js");
 
 // ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-function captureLogger(): { lines: string[]; logger: PluginLogger } {
-  const lines: string[] = [];
-  return {
-    lines,
-    logger: {
-      info: (msg: string) => lines.push(msg),
-      warn: (msg: string) => lines.push(`WARN: ${msg}`),
-      error: (msg: string) => lines.push(`ERROR: ${msg}`),
-      debug: (_msg: string) => {},
-    },
-  };
-}
-
-/** Create a mock spawn that emits events. */
-function mockSpawnProc(exitCode: number | null = 0, error?: Error) {
-  const handlers: Record<string, ((...args: unknown[]) => void)[]> = {};
-  const proc = {
-    on: vi.fn((event: string, handler: (...args: unknown[]) => void) => {
-      if (!handlers[event]) handlers[event] = [];
-      handlers[event].push(handler);
-      return proc;
-    }),
-    emit: (event: string, ...args: unknown[]) => {
-      for (const h of handlers[event] ?? []) h(...args);
-    },
-  };
-
-  vi.mocked(spawn).mockReturnValue(proc as never);
-
-  // Schedule events asynchronously so the promise in cliConnect can attach listeners
-  setTimeout(() => {
-    if (error) {
-      proc.emit("error", error);
-    } else {
-      proc.emit("close", exitCode);
-    }
-  }, 0);
-
-  return proc;
-}
-
-// ---------------------------------------------------------------------------
 // Tests
 // ---------------------------------------------------------------------------
 
@@ -71,7 +26,7 @@ beforeEach(() => {
 
 describe("cliConnect", () => {
   it("spawns openshell sandbox connect with correct sandbox name", async () => {
-    mockSpawnProc(0);
+    mockSpawnProc(vi.mocked(spawn), 0);
 
     const { lines, logger } = captureLogger();
     await cliConnect({ sandbox: "openclaw", logger });
@@ -87,7 +42,7 @@ describe("cliConnect", () => {
   });
 
   it("uses the provided sandbox name", async () => {
-    mockSpawnProc(0);
+    mockSpawnProc(vi.mocked(spawn), 0);
 
     const { logger } = captureLogger();
     await cliConnect({ sandbox: "my-custom-sandbox", logger });
@@ -100,7 +55,7 @@ describe("cliConnect", () => {
   });
 
   it("logs ENOENT error when openshell is not installed", async () => {
-    mockSpawnProc(null, new Error("ENOENT: spawn openshell not found"));
+    mockSpawnProc(vi.mocked(spawn), null, new Error("ENOENT: spawn openshell not found"));
 
     const { lines, logger } = captureLogger();
     await cliConnect({ sandbox: "openclaw", logger });
@@ -110,7 +65,7 @@ describe("cliConnect", () => {
   });
 
   it("logs generic error on non-ENOENT spawn failure", async () => {
-    mockSpawnProc(null, new Error("permission denied"));
+    mockSpawnProc(vi.mocked(spawn), null, new Error("permission denied"));
 
     const { lines, logger } = captureLogger();
     await cliConnect({ sandbox: "openclaw", logger });
@@ -120,7 +75,7 @@ describe("cliConnect", () => {
   });
 
   it("logs exit code error on non-zero exit", async () => {
-    mockSpawnProc(127);
+    mockSpawnProc(vi.mocked(spawn), 127);
 
     const { lines, logger } = captureLogger();
     await cliConnect({ sandbox: "openclaw", logger });
@@ -131,7 +86,7 @@ describe("cliConnect", () => {
   });
 
   it("does not log error on successful exit (code 0)", async () => {
-    mockSpawnProc(0);
+    mockSpawnProc(vi.mocked(spawn), 0);
 
     const { lines, logger } = captureLogger();
     await cliConnect({ sandbox: "openclaw", logger });
@@ -142,7 +97,7 @@ describe("cliConnect", () => {
   });
 
   it("does not log error on null exit code", async () => {
-    mockSpawnProc(null);
+    mockSpawnProc(vi.mocked(spawn), null);
 
     const { lines, logger } = captureLogger();
     await cliConnect({ sandbox: "openclaw", logger });
